Refactoring PyNvCodec module
