+++ PyNvCodec/src/PyFFMpegDecoder.cpp
@@ -1,0 +1,102 @@
+/*
+ * Copyright 2019 NVIDIA Corporation
+ * Copyright 2021 Kognia Sports Intelligence
+ * Copyright 2021 Videonetics Technology Private Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "PyNvCodec.hpp"
+
+using namespace std;
+using namespace VPF;
+using namespace chrono;
+
+namespace py = pybind11;
+
+constexpr auto TASK_EXEC_SUCCESS = TaskExecStatus::TASK_EXEC_SUCCESS;
+constexpr auto TASK_EXEC_FAIL = TaskExecStatus::TASK_EXEC_FAIL;
+
+PyFfmpegDecoder::PyFfmpegDecoder(const string& pathToFile,
+                                 const map<string, string>& ffmpeg_options)
+{
+  NvDecoderClInterface cli_iface(ffmpeg_options);
+  upDecoder.reset(FfmpegDecodeFrame::Make(pathToFile.c_str(), cli_iface));
+}
+
+bool PyFfmpegDecoder::DecodeSingleFrame(py::array_t<uint8_t>& frame)
+{
+  if (TASK_EXEC_SUCCESS == upDecoder->Execute()) {
+    auto pRawFrame = (Buffer*)upDecoder->GetOutput(0U);
+    if (pRawFrame) {
+      auto const frame_size = pRawFrame->GetRawMemSize();
+      if (frame_size != frame.size()) {
+        frame.resize({frame_size}, false);
+      }
+
+      memcpy(frame.mutable_data(), pRawFrame->GetRawMemPtr(), frame_size);
+      return true;
+    }
+  }
+  return false;
+}
+
+void* PyFfmpegDecoder::GetSideData(AVFrameSideDataType data_type,
+                                   size_t& raw_size)
+{
+  if (TASK_EXEC_SUCCESS == upDecoder->GetSideData(data_type)) {
+    auto pSideData = (Buffer*)upDecoder->GetOutput(1U);
+    if (pSideData) {
+      raw_size = pSideData->GetRawMemSize();
+      return pSideData->GetDataAs<void>();
+    }
+  }
+  return nullptr;
+}
+
+py::array_t<MotionVector> PyFfmpegDecoder::GetMotionVectors()
+{
+  size_t size = 0U;
+  auto ptr = (AVMotionVector*)GetSideData(AV_FRAME_DATA_MOTION_VECTORS, size);
+  size /= sizeof(*ptr);
+
+  if (ptr && size) {
+    py::array_t<MotionVector> mv({size});
+    auto req = mv.request(true);
+    auto mvc = static_cast<MotionVector*>(req.ptr);
+
+    for (auto i = 0; i < req.shape[0]; i++) {
+      mvc[i].source = ptr[i].source;
+      mvc[i].w = ptr[i].w;
+      mvc[i].h = ptr[i].h;
+      mvc[i].src_x = ptr[i].src_x;
+      mvc[i].src_y = ptr[i].src_y;
+      mvc[i].dst_x = ptr[i].dst_x;
+      mvc[i].dst_y = ptr[i].dst_y;
+      mvc[i].motion_x = ptr[i].motion_x;
+      mvc[i].motion_y = ptr[i].motion_y;
+      mvc[i].motion_scale = ptr[i].motion_scale;
+    }
+
+    return move(mv);
+  }
+
+  return move(py::array_t<MotionVector>({0}));
+}
+
+void Init_PyFFMpegDecoder(py::module& m)
+{
+  py::class_<PyFfmpegDecoder>(m, "PyFfmpegDecoder")
+      .def(py::init<const string&, const map<string, string>&>())
+      .def("DecodeSingleFrame", &PyFfmpegDecoder::DecodeSingleFrame)
+      .def("GetMotionVectors", &PyFfmpegDecoder::GetMotionVectors,
+           py::return_value_policy::move);
+}
